Move project header to its own component
